Fix NFT image and video download issue
