+++ src/containers/other/NFTView/hooks/useAsset.ts
@@ -1,5 +1,5 @@
 //- React Imports
-import { useMemo, useCallback } from 'react';
+import { useCallback } from 'react';
 
 //- Library Imports
 import { useCurrentDomain } from 'lib/providers/CurrentDomainProvider';
@@ -17,37 +17,29 @@ import {
 
 //- Hook level type definitions
 interface useAssetsReturn {
-	downloadAsset: () => Promise<void>;
+	downloadAsset: (assetUrl: string) => Promise<void>;
 	shareAsset: () => Promise<void>;
 }
 
 export const useAsset = (): useAssetsReturn => {
 	//- Current domain
-	const { domain: znsDomain, domainRaw: domain } = useCurrentDomain();
+	const { domainRaw: domain } = useCurrentDomain();
 
 	//- Notification
 	const { addNotification } = useNotification();
 
-	//- Memoized data
-	const domainAssetURL = useMemo(() => {
-		return (
-			znsDomain?.animation_url || znsDomain?.image_full || znsDomain?.image
-		);
-	}, [znsDomain]);
-
 	//- Callback functions
 
-	const downloadAsset = useCallback(async () => {
-		if (!domainAssetURL) {
-			return;
-		}
+	const downloadAsset = useCallback(
+		async (assetUrl: string) => {
+			const asset = await getDomainAsset(assetUrl);
 
-		try {
-			const asset = await getDomainAsset(domainAssetURL);
 			if (asset) {
-				addNotification(MESSAGES.DOWNLOAD, 1000);
 				try {
+					addNotification(MESSAGES.DOWNLOAD, 1000);
+
 					await downloadDomainAsset(asset);
+
 					// Set timeout to prevent overlapping notifications
 					setTimeout(() => {
 						addNotification(MESSAGES.DOWNLOAD_SUCCESSFUL);
@@ -57,11 +49,9 @@ export const useAsset = (): useAssetsReturn => {
 					addNotification(MESSAGES.DOWNLOAD_ERROR);
 				}
 			}
-		} catch (e) {
-			console.error(e);
-			addNotification(MESSAGES.ASSET_ERROR);
-		}
-	}, [domainAssetURL, addNotification]);
+		},
+		[addNotification],
+	);
 
 	//- Opens share
 	const shareAsset = useCallback(async () => {
